backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
passlib[bcrypt]
pyjwt
python-multipart
pandas
scikit-learn
yfinance
psycopg2-binary
